+++ Rest/AgentsRest/AgentsRest/Models/MissionModel.cs
@@ -12,7 +12,7 @@ namespace AgentsRest.Models
         public AgentModel Agent { get; set; }
         public int TargetId { get; set; }
         public TargetModel Target { get; set; }
-        public int TimeRemaining { get; set; }
+        public double TimeRemaining { get; set; }
         public int ActualExecutionTime { get; set; }
 
         public StatusMission StatusMission { get; set; }
